Refactor transaction broadcasting
The app was relying on the insight server to return a recently broadcasted
transaction over websockets so it could be ingested into the wallet through
the wallet service. If the server is being flakey, however, this means the
transaction would be lost.

This commit refactors the broadcast functionality to first send the transaction
to the wallet service to be ingested first before sending to the insight server.

It also removes the check where transactions not returned by the insight server
are deleted from the database. This way if the server is malfunctioning we are
not deleting unbroadcasted transactions.

Finally on each block if we have unconfirmed transactions in the db we will
rebroadcast.