Add client/errors pattern for responding to Retryable and Fatal errors
client/errors are consumed by client/pool to determine the appropriate
action when handling errors from the underlying client. The
wrappedErrors Fatal and Retryable are mutually exclusive and composable.
They are not intended to be used outside of the client package for
concerns of a wrappedError becoming unknowingly unwrapped.

Note that a wrappedError is unwrapped by the common error idiom
fmt.Errorf("additional context: %s", err.Error()) as calling err.Error()
can only return string and sheds all wrapped context.