Fix zcash balance calculation
This commit fixes a bug calculating zcash balances when spending confirmed
change. There were two ways to fix it, either build a zcash-specifc calculation
function or trim the zcash transaction so that the util.CalcBalance function
can be used. We opt for the later so as to stick with code that we know has
been working well and also because parts of the transaction that are being trimmed
are not relevant to the balance calculation.

closes #91